+++ ArduPlane/GCS_Mavlink.h
@@ -59,7 +59,7 @@ private:
     uint8_t radio_in_rssi() const;
 
     float vfr_hud_airspeed() const override;
-    int16_t vfr_hud_throttle() const override;
+    float vfr_hud_throttle() const override;
     float vfr_hud_climbrate() const override;
     
 #if HAL_HIGH_LATENCY2_ENABLED
